Changed libft.h and ft_atoi

+++ libs/libft/ft_atoi.c
@@ -6,15 +6,15 @@
 /*   By: muayna <muayna@student.42istanbul.com.t    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/30 10:49:04 by muayna            #+#    #+#             */
-/*   Updated: 2025/05/31 12:35:24 by muayna           ###   ########.fr       */
+/*   Updated: 2025/10/02 16:53:28 by muayna           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_atoi(const char *nptr)
+long long	ft_atoi(const char *nptr)
 {
-	int	i;
-	int	countnp;
-	int	number;
+	long	long	i;
+	long long	countnp;
+	long long number;
 
 	number = 0;
 	countnp = 1;
